Rename TargetCourseSelector as DestinationCourseSelector
