+++ client/app/bundles/course/duplication/pages/Duplication/DestinationCourseSelector/NewCourseForm.jsx
