+++ client/app/bundles/course/duplication/pages/Duplication/DestinationCourseSelector/index.jsx
